// bekend/model/User.ts
export class User {
    userId?: number;
    username?: string;
    password?: string;
    securityQuestion?: string;
    securityAnswer?: string;
    firstName?: string;
    lastName?: string;
    gender?: string;
    address?: string;
    contactPhone?: string;
    email?: string;
    profileImage?: Blob;
    schoolType?: string;
    currentGrade?: number;
    
  }